pyproject.toml
[tool.poetry]
name = "read-a-script"
version = "0.1.0"
description = "Mac: Read a play script in various voices, to help with line learning."
authors = ["Neil Padgen <neil.padgen@gmail.com>"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.9"
readchar = "^2.0.1"
Jouvence = "^0.4.2"
docopt = "^0.6.2"
"ruamel.yaml" = "^0.16.12"
pyttsx3 = "^2.90"

[tool.poetry.dev-dependencies]
ipython = "^7.19.0"
jupyter = "^1.0.0"
dephell = "^0.8.3"

# Following is commented out because of a dephell error
# you must generate setup.py with the command
#  dephell deps convert --from-format=poetry --from-path=pyproject.toml --to-format=setuppy --to-path=setup.py
# [tool.dephell.main]
# from = {format = "poetry", path = "pyproject.toml"}
# to = {format = "setuppy", path = "setup.py"}

[tool.poetry.scripts]
script_learner = "read_a_script.script_learner:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
